pyproject.toml
[tool.poetry]
name = "ProjBase"
version = "0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.22.4"
scipy = "^1.8.1"
torch = "^1.11.0"
torchvision = "^0.12.0"
opencv-python = "^4.5.5"
imgaug = "^0.4.0"
PyYAML = "^6.0"
yacs = "^0.1.8"
transformers = "^4.19.2"
pandas = "^1.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
